Added dtos, changed controllers

+++ entity/src/main/java/ua/lviv/yuriizhurakovskyi/repository/ClassworkRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ua.lviv.yuriizhurakovskyi.entity.Classwork;
 
 @Repository
-public interface ClassworkRepository extends JpaRepository<Classwork, Integer> {
+public interface ClassworkRepository extends JpaRepository<Classwork, Long> {
 }
